requirements.txt
# Config + HTTP + parsing
PyYAML==6.0.2
requests==2.32.3
tenacity==9.0.0
feedparser==6.0.11
python-dateutil==2.9.0.post0

# Newsletter/article text extraction
trafilatura==1.9.0

# YouTube transcripts
youtube-transcript-api==0.6.2

# LLM calls
openai==1.45.0

# Local ASR fallback (Whisper)
faster-whisper==1.0.3
ctranslate2==4.4.0

# Stage 7 Google Sheets persistence
gspread==6.1.4
google-auth==2.35.0

# Testing
pytest==8.3.5
